refactor: fsd 구조에 맞게 수정
