+++ apps/web/domain-shared/auth/index.ts
@@ -1,0 +1,7 @@
+// API
+export * from "./api/authClient";
+export * from "./api/types";
+
+// Lib
+export * from "./lib/appleAuth";
+export * from "./lib/authUtils";
