--- apps/web/domain-shared/auth/lib/index.ts
@@ -1,6 +1,0 @@
-import { sayGood } from "@/shared/lib";
-
-export function sayHelloAndGood() {
-  console.log("Hello");
-  sayGood();
-}
