+++ apps/web/shared/lib/index.ts
@@ -1,5 +1,3 @@
 export function sayGood() {
   console.log("Good");
 }
-
-export * from "./appleAuth";
