+++ apps/web/shared/ui/Splash.tsx
@@ -14,16 +14,12 @@ enum SplashText {
   moveToLogin = "로그인 페이지로 이동 중...",
 }
 
-export function Splash({
-  type = "loading",
-  logoSize = { width: undefined, height: undefined },
-  className = "",
-}: SplashProps) {
+export function Splash({ type = "loading", logoSize, className = "" }: SplashProps) {
   return (
     <div className={`min-h-screen flex items-center justify-center bg-[#FF6B4C] ${className}`}>
       <div className="text-center text-white space-y-4">
-        <Logo width={logoSize.width} height={logoSize.height} />
-        <p className="text-lg">{SplashText[type]}</p>
+        <Logo width={logoSize?.width} height={logoSize?.height} />
+        <p className="mt-4 text-lg">{SplashText[type]}</p>
       </div>
     </div>
   );
