+++ apps/web/shared/ui/index.ts
@@ -1,2 +1,3 @@
 export * from "./Button";
 export * from "./Card";
+export * from "./Splash";
